Add code for the bidirectional watering pot
The code automatically fills the watering pot dependent on the moisture sensor (to 3/4 of the watreing pot). Dependent on the surrounding noise the watering pot will either water the one side (labeled as PERSON) or the other side (labeled as PLANT). In addition, the second moisture sensor checks the plants soil status and may water the plant.